lift add + remove interval functions to parent component
